Create the basic setup and add navbar and sections
Create the basic setup (index.html and style.css) to develop
the project. Add sections with id and classes, add a header
corresponding to each section. Add a navbar with id, add a
header, add links to each section with corresponding names.

The basic setup has all the files needed to develop the
project. The sections with its features are requirements
of the project. The navbar with its features are
requirements demanded of the project.