Add an image and finish the HTML, CSS coding
Add an image ilustrating  Git file state system. Finish the
HTML code adding all the required stuff. Finish the CSS
code complying all the requirements.

The image was added to help the reader understand the Git
file state system. The HTML code was added to comply with
the project requirements. The CSS code was added to
customize the page according to a documentation and to
conform to the project requirements.